+++ cli/util/path.rs
@@ -241,8 +241,6 @@ pub fn to_percent_decoded_str(s: &str) -> String {
 
 #[cfg(test)]
 mod test {
-  use deno_runtime::fs_util::specifier_to_file_path;
-
   use super::*;
 
   #[test]
@@ -285,24 +283,6 @@ mod test {
     assert!(is_importable_ext(Path::new("foo.cts")));
     assert!(is_importable_ext(Path::new("foo.json")));
     assert!(!is_importable_ext(Path::new("foo.mjsx")));
-  }
-
-  #[test]
-  fn test_specifier_to_file_path() {
-    run_success_test("file:///", "/");
-    run_success_test("file:///test", "/test");
-    run_success_test("file:///dir/test/test.txt", "/dir/test/test.txt");
-    run_success_test(
-      "file:///dir/test%20test/test.txt",
-      "/dir/test test/test.txt",
-    );
-
-    fn run_success_test(specifier: &str, expected_path: &str) {
-      let result =
-        specifier_to_file_path(&ModuleSpecifier::parse(specifier).unwrap())
-          .unwrap();
-      assert_eq!(result, PathBuf::from(expected_path));
-    }
   }
 
   #[test]
